Former-commit-id: f1c00eeaefb1309eaa5c09fa21d0bf58dfcd3cba

+++ recovery-coreextension-plugin/src/main/java/es/pfsgroup/plugin/recovery/nuevoModeloBienes/api/model/NMBAdicionalBienInfo.java
@@ -1,9 +1,9 @@
 package es.pfsgroup.plugin.recovery.nuevoModeloBienes.api.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import es.capgemini.pfs.auditoria.model.Auditoria;
-import es.capgemini.pfs.bien.model.DDTipoBien;
 import es.pfsgroup.plugin.recovery.nuevoModeloBienes.model.DDTipoInmueble;
 import es.pfsgroup.plugin.recovery.nuevoModeloBienes.model.DDTipoProdBancario;
 
@@ -34,7 +34,7 @@ public interface NMBAdicionalBienInfo {
 
 	DDTipoInmueble getTipoInmueble();
 
-	Float getValoracion();
+	BigDecimal getValoracion();
 
 	String getEntidad();
